+++ apps/mobile/src/features/AddressBook/index.tsx
@@ -1,3 +1,3 @@
-export { ListContainer } from './List/List.container'
-export { ContactContainer } from './Contact/Contact.container'
-export { ContactNameContainer } from './Contact/ContactName.container'
+export { AddressBookListContainer } from './List/AddressBookList.container'
+export { ContactDetailContainer } from './Contact/ContactDetail.container'
+export { ContactDisplayNameContainer } from './Contact/ContactDisplayName.container'
