+++ apps/mobile/src/features/ConfirmationsSheet/ConfirmationsSheet.container.tsx
@@ -14,7 +14,7 @@ import {
 } from '@safe-global/store/gateway/AUTO_GENERATED/transactions'
 import { selectSigners } from '@/src/store/signersSlice'
 import { useAppSelector } from '@/src/store/hooks'
-import { ContactNameContainer } from '../AddressBook'
+import { ContactDisplayNameContainer } from '../AddressBook'
 
 export const ConfirmationsSheetContainer = () => {
   const activeSafe = useDefinedActiveSafe()
@@ -61,7 +61,7 @@ export const ConfirmationsSheetContainer = () => {
       return (
         <View width="100%">
           <SignersCard
-            name={<ContactNameContainer address={item.value as Address} />}
+            name={<ContactDisplayNameContainer address={item.value as Address} />}
             getSignerTag={getSignerTag}
             address={item.value as Address}
             rightNode={
